requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.1.0
psycopg2==2.8.5